navigation_system.py - extracted the PID controller as its own class. Feel like we may be needing it again.
ship.py - Fixed the issue with multiple showbase being called. Ship now successfully extends Actor and will render the model that is passed to it.

pid.py - simple PID controller for flight control.

test.py - currently the main class. It loads a ship, renders it to the screen, and moves it around to several locations.